fp/requirements.txt
eventlet==0.33.3
pygame==2.4.0
python-socketio==5.8.0
